requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.28.44
botocore==1.31.44
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==2.0
django-js-asset==1.2.2
django-recaptcha==3.0.0
django-storages==1.13.2
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==1.0.1
oauthlib==3.1.1
Pillow==9.0.1
psycopg2-binary==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.4.0
PyMySQL==1.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.6.2
service-identity==21.1.0
six==1.16.0
smsir-python==1.0.7
sqlparse==0.4.2
Twisted==22.4.0
txaio==21.2.1
typing-extensions==4.7.1
tzdata==2022.1
urllib3==1.26.6
zope.interface==5.4.0
